requirements.txt
annotated-types==0.7.0
Faker==38.2.0
kafka-python==2.3.0
prometheus-client==0.21.0
pydantic==2.12.5
pydantic_core==2.41.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
pytest
pre-commit
